refactor: clean up secondary cache interface
